+++ R/print.R
@@ -11,9 +11,7 @@
 
 print.gh_response <- function(x, ...) {
   if (inherits(x, c("raw", "path"))) {
-    attr(x, c("method")) <- NULL
-    attr(x, c("response")) <- NULL
-    attr(x, ".send_headers") <- NULL
+    attributes(x) <- list(class = class(x))
     print.default(x)
   } else {
     print(toJSON(unclass(x), pretty = TRUE, auto_unbox = TRUE, force = TRUE))